// blockchain/typechain-types/VoterRegistry.ts
export declare namespace VoterRegistry {
  export type VoterStruct = {
    nim: string;
    name: string;
    role: BigNumberish;
    isRegistered: boolean;
    registeredAt: BigNumberish;
  };

  export type VoterStructOutput = [
    nim: string,
    name: string,
    role: bigint,
    isRegistered: boolean,
    registeredAt: bigint
  ] & {
    nim: string;
    name: string;
    role: bigint;
    isRegistered: boolean;
    registeredAt: bigint;
  };
}