// blockchain/typechain-types/VeriVoteMain.ts
export declare namespace VeriVoteMain {
  export type CandidateStruct = {
    id: BigNumberish;
    name: string;
    description: string;
    imageHash: string;
    voteCount: BigNumberish;
  };

  export type CandidateStructOutput = [
    id: bigint,
    name: string,
    description: string,
    imageHash: string,
    voteCount: bigint
  ] & {
    id: bigint;
    name: string;
    description: string;
    imageHash: string;
    voteCount: bigint;
  };

  export type ElectionStruct = {
    id: BigNumberish;
    title: string;
    description: string;
    status: BigNumberish;
    createdBy: AddressLike;
    createdAt: BigNumberish;
    startDate: BigNumberish;
    endDate: BigNumberish;
    totalVoters: BigNumberish;
    totalVotes: BigNumberish;
    candidateCount: BigNumberish;
  };

  export type ElectionStructOutput = [
    id: bigint,
    title: string,
    description: string,
    status: bigint,
    createdBy: string,
    createdAt: bigint,
    startDate: bigint,
    endDate: bigint,
    totalVoters: bigint,
    totalVotes: bigint,
    candidateCount: bigint
  ] & {
    id: bigint;
    title: string;
    description: string;
    status: bigint;
    createdBy: string;
    createdAt: bigint;
    startDate: bigint;
    endDate: bigint;
    totalVoters: bigint;
    totalVotes: bigint;
    candidateCount: bigint;
  };
}